+++ src/format.js
@@ -1,6 +1,20 @@
-import {splitHead, unquote} from './util'
+export function splitHead (str, sep) {
+  const idx = str.indexOf(sep)
+  if (idx === -1) return [str]
+  return [str.slice(0, idx), str.slice(idx + sep.length)]
+}
 
-export default function format (nodes) {
+export function unquote (str) {
+  const car = str.charAt(0)
+  const end = str.length - 1
+  const isQuoteStart = car === '"' || car === "'"
+  if (isQuoteStart && car === str.charAt(end)) {
+    return str.slice(1, end)
+  }
+  return str
+}
+
+export function format (nodes) {
   return nodes.map(node => {
     const type = node.type
     if (type === 'element') {
