--- translate.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/translate/v0').default
